requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot[job-queue]
python-multipart
aiofiles
jinja2
pydantic-settings
sqlalchemy
aiosqlite
httpx
python-dotenv
telethon
